+++ packages/babel-plugin-transform-es3-member-expression-literals/src/index.js
@@ -4,7 +4,7 @@ export default function ({ types: t }) {
       MemberExpression: {
         exit({ node }) {
           let prop = node.property;
-          if (!node.computed && t.isIdentifier(prop) && !t.isValidIdentifier(prop.name)) {
+          if (!node.computed && t.isIdentifier(prop) && !t.isSpecIdentifier(prop.name)) {
             // foo.default -> foo["default"]
             node.property = t.stringLiteral(prop.name);
             node.computed = true;