+++ packages/babel-plugin-transform-property-literals/src/index.js
@@ -4,7 +4,7 @@ export default function ({ types: t }) {
       ObjectProperty: {
         exit({ node }) {
           let key = node.key;
-          if (t.isLiteral(key) && t.isValidIdentifier(key.value)) {
+          if (t.isLiteral(key) && t.isSpecIdentifier(key.value)) {
             // "foo": "bar" -> foo: "bar"
             node.key = t.identifier(key.value);
             node.computed = false;